fix: إصلاح أخطاء نوع Decimal في العمليات الحسابية
✅ إضافة .toDecimal() لجميع العمليات الحسابية على Decimal
✅ إصلاح database_helper.dart في getPartnerAvailableBalance
✅ إصلاح supplier_details_report_screen.dart في 3 أماكن

المشكلة: العمليات على Decimal تُرجع Rational وليس Decimal
الحل: استخدام .toDecimal() للتحويل الصحيح

+++ lib/data/database_helper.dart
@@ -1778,13 +1778,13 @@ Future<Decimal> getTotalDeductions() async {
     final netProfit = totalProfit - totalWithdrawnForSupplier;
 
     // 2️⃣ حساب نصيب هذا الشريك من صافي الربح
-    final partnerTotalShare = (netProfit * sharePercentage / Decimal.fromInt(100));
+    final partnerTotalShare = (netProfit * sharePercentage / Decimal.fromInt(100)).toDecimal();
 
     // 3️⃣ حساب كم سحب هذا الشريك بالفعل
     final partnerWithdrawn = await getTotalWithdrawnForPartner(supplierId, partnerName);
 
     // 4️⃣ الرصيد المتاح = نصيبه - ما سحبه
-    final availableBalance = partnerTotalShare - partnerWithdrawn;
+    final availableBalance = (partnerTotalShare - partnerWithdrawn).toDecimal();
 
     return availableBalance;
   }
